Modify install-chef to look for a null realm instead of "Crowbar" in the default template.

+++ redhat-5.6-extra/install-chef.sh
@@ -144,20 +144,23 @@ cat /root/.ssh/id_rsa.pub >>/root/.ssh/authorized_keys
 cp /root/.ssh/authorized_keys \
     /opt/dell/chef/cookbooks/redhat-install/files/default/authorized_keys
 
+CROWBAR_REALM=""
 # generate the machine install username and password
-if [[ ! -e /etc/crowbar.install.key ]]; then
+if [[ ! -e /etc/crowbar.install.key && $CROWBAR_REALM]]; then
     dd if=/dev/urandom bs=65536 count=1 2>/dev/null |sha512sum - 2>/dev/null | \
 	(read key rest; echo "machine-install:$key" >/etc/crowbar.install.key)
     export CROWBAR_KEY=$(cat /etc/crowbar.install.key)
-    printf "${CROWBAR_KEY%%:*}:Crowbar:${CROWBAR_KEY##*:}" |md5sum - | (
-	read key rest
-	printf "\n${CROWBAR_KEY%%:*}:Crowbar:$key\n" >> \
+    printf "${CROWBAR_KEY%%:*}:${CROWBAR_REALM}:${CROWBAR_KEY##*:}" | \
+	md5sum - | (read key rest
+	printf "\n${CROWBAR_KEY%%:*}:${CROWBAR_REALM}:$key\n" >> \
 	    /opt/dell/openstack_manager/htdigest)
-else
-  export CROWBAR_KEY=$(cat /etc/crowbar.install.key)
-fi
-sed -i "s/machine_password/${CROWBAR_KEY##*:}/g" \
+elif [[ $CROWBAR_REALM ]]
+    export CROWBAR_KEY=$(cat /etc/crowbar.install.key)
+    sed -i -e "s/machine_password/${CROWBAR_KEY##*:}/g" \
+	-e "/\"realm\":/ s/null/\"$CROWBAR_REALM\"/g"
     /opt/dell/chef/data_bags/crowbar/bc-template-crowbar.json
+fi
+    
 
 # Crowbar will hack up the pxeboot files appropriatly.
 # Set Version in Crowbar UI
